Metadata.get() was mis-typed.  The type was incorrectly Union[None, T1, T2], where T1 is the type of $default and T2 is the type of $cls. Therefore, if T1==T2, you'd get an Optional[T1] instead of a T1.
That meant you couldn't write something like this:
md = Metadata()
md['x'] = Duration(sec=60)
md.get('x', Duration(sec=0)).sec

because the type checker would think that the md.get(...) call has type
Union[Duration, None] and it doesn't allow the ".sec" attibute to be taken
because None doesn't have it.  (It's trying to guarantee that you'll never get
an attribute error.)

The fix is to tweak the typing logic.

But it's also sometimes convenient for the code to raise an exception,
so that the user can write:

md.get('x').sec

instead of

tmp = md.get('x')
if x is not None:
  tmp.sec

so I added the get_or_error() method as an analog for  md['key', cls=Duration]
which would be nice, but seems syntactically impossible.

PiperOrigin-RevId: 466716628