Adds the `pe_overwrite_probability` parameter to the `UCBPEConfig`
This parameter configures the probability with which the `VizierGPUCBPEBandit` designer chooses the pure exploration acquisition function when there are newly completed trials. The default value is 0.0, which is a no-op from the behavior of the base code.

In the fully sequential use case where every suggestion is completed before the next one is requested, setting this parameter to a positive value `p` in [0, 1] configures the `VizierGPUCBPEBandit` designer to perform (controlled) exploration on roughly `p`-fraction of the suggestions.

PiperOrigin-RevId: 627101720